Autorisation functionnality and bug fixing are done

+++ src/Controller/TaskController.php
@@ -3,11 +3,14 @@
 namespace App\Controller;
 
 use App\Entity\Task;
+use App\Entity\User;
 use App\Form\TaskType;
 use App\Repository\TaskRepository;
+use App\Repository\UserRepository;
+use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\Routing\Annotation\Route;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
-use Symfony\Component\HttpFoundation\Request;
+use Symfony\Component\Security\Core\Exception\AccessDeniedException;
 
 class TaskController extends AbstractController
 {
@@ -22,7 +25,7 @@ class TaskController extends AbstractController
     /**
      * @Route("/tasks/create", name="task_create")
      */
-    public function createAction(Request $request)
+    public function createAction(TaskRepository $taskRepository, UserRepository $userRepository, Request $request)
     {
         $task = new Task();
         $form = $this->createForm(TaskType::class, $task);
@@ -30,6 +33,15 @@ class TaskController extends AbstractController
         $form->handleRequest($request);
 
         if ($form->isSubmitted() && $form->isValid()) {
+            if($this->getUser() == Null)
+            {
+                $taskRepository->addAnonymousUser($task);
+
+            }else
+            {
+                $task->setUser($this->getUser());
+            }
+            
             $em = $this->getDoctrine()->getManager();
 
             $em->persist($task);
@@ -46,8 +58,10 @@ class TaskController extends AbstractController
     /**
      * @Route("/tasks/{id}/edit", name="task_edit")
      */
-    public function editAction(Task $task, Request $request)
+    public function editAction($id, UserRepository $userRepository, TaskRepository $taskRepository, Request $request)
     {
+        $task = $taskRepository->find($id);
+        $task = $taskRepository->addAnonymousUserIfNeeded($task);
         $form = $this->createForm(TaskType::class, $task);
 
         $form->handleRequest($request);
@@ -83,14 +97,28 @@ class TaskController extends AbstractController
     /**
      * @Route("/tasks/{id}/delete", name="task_delete")
      */
-    public function deleteTaskAction($id, TaskRepository $taskRepository)
+    public function deleteTaskAction($id,UserRepository $userRepository, TaskRepository $taskRepository, Request $request)
     {
         $task = $taskRepository->find($id);
-        $em = $this->getDoctrine()->getManager();
-        $em->remove($task);
-        $em->flush();
+        if (
+            ($task->getUser() == $this->getUser()) or
+            (
+                (
+                    ( $task->getUser() == null ) or
+                    ( $task->getUser() == $userRepository->findOneByUsername('Anonyme') )
+                ) and
+                $this->isGranted('ROLE_ADMIN')
+            )
+           )
+        {
+            $em = $this->getDoctrine()->getManager();
+            $em->remove($task);
+            $em->flush();
+            $this->addFlash('success', 'La tâche a bien été supprimée.');
+        }else{
+            $request->getSession()->getFlashBag()->add('error', 'Vous ne pouvez pas supprimer une tâche créé par un autre utilisateur.');
+        }
 
-        $this->addFlash('success', 'La tâche a bien été supprimée.');
 
         return $this->redirectToRoute('task_list');
     }
